+++ cp/src/main/java/cp/nsga/operator/mutation/melody/OneNoteMutation.java
@@ -64,7 +64,7 @@ public class OneNoteMutation extends AbstractMutation {
 			String name = cls.getName();
 			throw new JMException("Exception in " + name + ".execute()");
 		}
-		doMutation(probability.doubleValue(), solution);
+		doMutation(probability, solution);
 		return solution;
 	}
