+++ cp/src/main/java/cp/nsga/operator/mutation/melody/ReplaceMelody.java
@@ -109,7 +109,7 @@ public class ReplaceMelody extends AbstractMutation{
 			String name = cls.getName();
 			throw new JMException("Exception in " + name + ".execute()");
 		}
-		doMutation(probability.doubleValue(), solution);
+		doMutation(probability, solution);
 		return solution;
 	}
